removed both assistantUtils.jsx and separate out into several utils/*.js* files
