+++ src/components/NavItems/Assistant/AssistantScrapeResults/AssistantTextClassification.jsx
@@ -27,7 +27,7 @@ import {
   rgbToString,
   treeMapToElements,
   wrapPlainTextSpan,
-} from "@/components/NavItems/Assistant/AssistantScrapeResults/assistantUtils";
+} from "@/components/NavItems/Assistant/utils";
 import { i18nLoadNamespace } from "@/components/Shared/Languages/i18nLoadNamespace";
 import useMyStyles from "@/components/Shared/MaterialUiStyles/useMyStyles";
 import { setImportantSentenceThreshold } from "@/redux/actions/tools/assistantActions";
