+++ src/components/NavItems/Assistant/utils/index.js
@@ -1,0 +1,31 @@
+export {
+  interpRgb,
+  rgbToString,
+  rgbToLuminance,
+  rgbListToGradient,
+  getMgtColours,
+  getSubjectivityColours,
+} from "./colorUtils";
+
+export {
+  renderSourceTypeChip,
+  prependHttps,
+  getUrlTypeFromCredScope,
+  renderThisDomainOrAccount,
+  renderScope,
+  renderLabels,
+  renderDescription,
+  renderEvidence,
+  renderDomainTitle,
+  renderDomainAnalysisResults,
+} from "./domainUtils";
+
+export {
+  treeMapToElements,
+  wrapPlainTextSpan,
+  mergeSpanIndices,
+  getPersuasionCategoryTechnique,
+  createGaugeChart,
+  ThresholdSlider,
+  scrollToElement,
+} from "./textUtils";
