+++ src/components/NavItems/Assistant/utils/textUtils.jsx
@@ -11,78 +11,6 @@ import _ from "lodash";
 import { v4 as uuidv4 } from "uuid";
 
 /**
- * Interpolate RGB between an arbitrary range
- * @param value
- * @param low
- * @param high
- * @param rgbLow RGB Value represented as an array e.g. [255, 255, 255] for white
- * @param rgbHigh RGB Value represented as an array e.g. [255, 255, 255] for white
- * @returns {*[]}
- */
-export const interpRgb = (value, low, high, rgbLow, rgbHigh) => {
-  let interp = value;
-  if (value < low) interp = low;
-  if (value > high) interp = high;
-  interp = (interp - low) / (high - low);
-
-  let output = [];
-  for (let i = 0; i < rgbLow.length; i++) {
-    let channelLow = rgbLow[i];
-    let channelHigh = rgbHigh[i];
-    let delta = channelHigh - channelLow;
-    output.push(channelLow + delta * interp);
-  }
-
-  return output;
-};
-
-/**
- * Converts an array-based RGB representation to rgba() format used by CSS
- * @param rgb RGB value represented as an array e.g. [255, 255, 255] for white
- * @returns {string}
- */
-export const rgbToString = (rgb) => {
-  return "rgba(" + rgb[0] + "," + rgb[1] + "," + rgb[2] + ",1)";
-};
-
-/**
- * Calculates the luminance of an RGB colour
- * @param rgb RGB value represented as an array e.g. [255, 255, 255] for white
- * @returns {number}
- */
-export const rgbToLuminance = (rgb) => {
-  let rgbNorm = rgb.map((c) => c / 255);
-  let rgbLinear = rgbNorm.map((c) => {
-    if (c <= 0.04045) {
-      return c / 12.92;
-    } else {
-      return Math.pow((c + 0.055) / 1.055, 2.4);
-    }
-  });
-  let luminance =
-    rgbLinear[0] * 0.2126 + rgbLinear[1] * 0.7152 + rgbLinear[2] * 0.0722;
-  return luminance;
-};
-
-/**
- * Generate CSS gradient from a list of RGB values
- * @param rgbList List RGB value represented as an array e.g. [[0,0,0], [255, 255, 255]] for a black to white gradient
- * @returns {string}
- */
-export const rgbListToGradient = (rgbList) => {
-  let gradientStr = "";
-  for (let i = 0; i < rgbList.length; i++) {
-    let colourStr = rgbToString(rgbList[i]);
-    let percentageStr = Math.round((i / (rgbList.length - 1)) * 100).toString();
-
-    if (i > 0) gradientStr += ",";
-    gradientStr += colourStr + " " + percentageStr + "%";
-  }
-  const output = "linear-gradient(90deg, " + gradientStr + ")";
-  return output;
-};
-
-/**
  * Recursively takes tree map and returns highlighted spans
  * @param text
  * @param treeElem
@@ -202,7 +130,7 @@ export const treeMapToElements = (
 };
 
 /**
- * Function for wrapping hightlighted spans
+ * Function for wrapping highlighted spans
  * @param text
  * @param spanHighlightIndices
  * @param wrapFunc
@@ -250,8 +178,6 @@ export const wrapPlainTextSpan = (text, spanHighlightIndices, wrapFunc) => {
  * @returns {*[]}
  */
 export const mergeSpanIndices = (filteredClassification) => {
-  // classification variable is a map of categories where each one has a list of classified spans, we
-  // have to invert that so that we have a list of spans that contains all categories in that span
   let mergedSpanIndices = [];
   for (let label in filteredClassification) {
     for (let i = 0; i < filteredClassification[label].length; i++) {
@@ -287,7 +213,9 @@ export const mergeSpanIndices = (filteredClassification) => {
   return mergedSpanIndices;
 };
 
-// persuasion techniques: split text for category and technique
+/**
+ * Persuasion techniques: split text for category and technique
+ */
 export function getPersuasionCategoryTechnique(category) {
   return category.split("__");
 }
@@ -347,46 +275,6 @@ export function createGaugeChart(
       </Box>
     </>
   );
-}
-
-/**
- * Defines the colours required by machine generated text
- * @param configs
- * @returns colours for light and dark mode
- */
-export function getMgtColours(configs) {
-  const colours = [
-    rgbToString(configs.greenRgb),
-    rgbToString(configs.lightGreenRgb),
-    rgbToString(configs.orangeRgb),
-    rgbToString(configs.redRgb),
-  ];
-  const coloursDark = [
-    rgbToString(configs.greenRgbDark),
-    rgbToString(configs.lightGreenRgbDark),
-    rgbToString(configs.orangeRgbDark),
-    rgbToString(configs.redRgbDark),
-  ];
-  return [colours, coloursDark];
-}
-
-/**
- * Defines the colours required by subjectivity
- * @param configs
- * @returns colours for light and dark mode
- */
-export function getSubjectivityColours(configs) {
-  const colours = [
-    rgbToString(configs.greenRgb),
-    rgbToString(configs.orangeRgb),
-    rgbToString(configs.redRgb),
-  ];
-  const coloursDark = [
-    rgbToString(configs.greenRgbDark),
-    rgbToString(configs.orangeRgbDark),
-    rgbToString(configs.redRgbDark),
-  ];
-  return [colours, coloursDark];
 }
 
 /**
